Split core.py
